+++ .gitignore
@@ -1,6 +1,6 @@
 /target/
 !.mvn/wrapper/maven-wrapper.jar
-naruto.iml
+naruto/naruto.iml
 
 ### STS ###
 .apt_generated
@@ -193,3 +193,5 @@ README.html
 *.iml
 .idea
 .exercism
+/gradle/
+/src/naruto/out/production/naruto/
