--- naruto/src/interfaces/NinjaInterface.java
@@ -1,7 +1,0 @@
-package interfaces;
-
-public interface NinjaInterface {
-
-    String usarJutsu();
-    String desviar(String desviar);
-}
