+++ src/naruto/src/interfaces/NinjaInterface.java
@@ -1,0 +1,7 @@
+package naruto.src.interfaces;
+
+public interface NinjaInterface {
+
+    String usarJutsu();
+    String desviar(String desviar);
+}
